Update Plugins
plugin.json files will now include a "version" field.
When the program is run, the installed field will be compared with the remote version and the user will be notified if it has changed

README.md has been changed to reflect this new field

Fix issue with organizing
I used [] to search the fields for urls in organizer.py If the field does not exist that would abort the organization of the file. Changed to use .get() because not having a comment field should not be fatal